Fix task 8

+++ flag.c
@@ -43,7 +43,7 @@ int *size, char **buf, int len_buf, const char **f, fs **fos)
 				if (check_c('l', i, *f - 1 - m) == 0)
 					j = va_arg(*ar, unsigned long int);
 				else if (check_c('h', i, *f - 1 - m) == 0)
-					j = va_arg(*ar, unsigned short int);
+					j = (unsigned short int)va_arg(*ar, int);
 				else
 					j = va_arg(*ar, unsigned int);
                                 if (j > 0)
@@ -63,7 +63,7 @@ int *size, char **buf, int len_buf, const char **f, fs **fos)
 				if (check_c('l', i, *f - 1 - m) == 0)
 					j = va_arg(*ar, unsigned long int);
 				else if (check_c('h', i, *f - 1 - m) == 0)
-					j = va_arg(*ar, unsigned short int);
+					j = (unsigned short int)va_arg(*ar, int);
 				else
 					j = va_arg(*ar, unsigned int);
 				if (j > 0)
@@ -83,7 +83,7 @@ int *size, char **buf, int len_buf, const char **f, fs **fos)
 				if (check_c('l', i, *f - 1 - m) == 0)
                                         j = va_arg(*ar, unsigned long int);
                                 else if (check_c('h', i, *f - 1 - m) == 0)
-                                        j = va_arg(*ar, unsigned short int);
+                                        j = (unsigned short int)va_arg(*ar, int);
                                 else
                                         j = va_arg(*ar, unsigned int);
 				if (j > 0)
@@ -108,7 +108,7 @@ int *size, char **buf, int len_buf, const char **f, fs **fos)
 				if (check_c('l', i, *f - 1 - m) == 0)
                                         j = va_arg(*ar, long int);
                                 else if (check_c('h', i, *f - 1 - m) == 0)
-                                        j = va_arg(*ar, short int);
+                                        j = (short int)va_arg(*ar, int);
                                 else
                                         j = va_arg(*ar, int);
 				if (j > -1)
@@ -133,7 +133,7 @@ int *size, char **buf, int len_buf, const char **f, fs **fos)
                                 if (check_c('l', i, *f - 1 - m) == 0)
                                         j = va_arg(*ar, long int);
                                 else if (check_c('h', i, *f - 1 - m) == 0)
-                                        j = va_arg(*ar, short int);
+                                        j = (short int)va_arg(*ar, int);
                                 else
                                         j = va_arg(*ar, int);
                                 if (j > -1)
